orchestration/airflow/requirements.txt
requests
pyspark==4.0.1
pytest
responses
apache-airflow==2.10.0
apache-airflow-providers-cncf-kubernetes